+++ com/uart.c
@@ -1,4 +1,4 @@
-const int tx_pin = PIN2_bm; // Port B
+const uint8_t tx_pin = PIN2_bm; // Port B
 const uint16_t uart_baudrate = 1388; //Specific to 3.33MHz
 
 void uart_init (){
